update the order cancel service to handle retry cases.

+++ src/modules/orders/orders-cancel.service.ts
@@ -17,6 +17,7 @@ import { LoggingService } from '../../services/logging.service';
 import { ScheduleMapping } from 'src/entities/schedule_mapping.entity';
 import { OrchestratorService } from '../orchestrator/orchestrator.service';
 import { MOVE_TYPE } from 'src/entities/task.entity';
+import { retry } from 'rxjs';
 
 
 
@@ -52,7 +53,9 @@ export class OrdersCancelService {
     if (task && (task.status === TaskStatus.CANCELLED || task.status === TaskStatus.COMPLETED || task.status === TaskStatus.PROCESSING)){
       throw new BadRequestException(`Cannot cancel tasks with status ${task.status}`);
     }
-    await this.productRequirementRepository.delete({ source_location_id: orderItem.source_location_id });
+    if (orderItem.retry_reassign_attempts >= 1){
+      await this.productRequirementRepository.delete({ source_location_id: orderItem.source_location_id });
+    }
     if (!task){return;}
     await this.orchestrationService.decrementRobotInUse();
     await this.loggingService.log(`Cancelling Task ID ${task.task_id} related to Order Item ID ${orderItem.order_item_id}`,
@@ -121,19 +124,95 @@ export class OrdersCancelService {
         status: In([OrderItemStatus.IN_PROGRESS])
       },
     });
+    let any_retry_exceed = false;
     for (const orderItem of orderItems){
+      if (orderItem.retry_reassign_attempts >= 1){
+        any_retry_exceed = true;
+      }
       orderItem.status = OrderItemStatus.CANCELLED;
       await this.orderItemRepository.save(orderItem);
       await this.loggingService.log(`Order Item ID ${orderItem.order_item_id} cancelled (via task ID ${taskId})`,
         TaskType.GOODS_TO_PERSON, null, orderItem.order_batch_id || '');
     }
-    await this.orchestrationService.decrementRobotInUse();
-    await this.productRequirementRepository.delete({ source_location_id: source_location_id });
+    // await this.orchestrationService.decrementRobotInUse();
+    if (any_retry_exceed){
+      await this.productRequirementRepository.delete({ source_location_id: source_location_id });
+    }
+    // await this.productRequirementRepository.delete({ source_location_id: source_location_id });
     await this.orchestrationService.CancelTask(task);
 
   }
   
   async retryOrderByTaskId(taskId: string){
+    const task = await this.taskRepository.findOne({ where: { task_id: taskId } });
+    if (!task){
+      throw new NotFoundException(`Task with ID ${taskId} not found`);
+    }
+    if (task.move_type === MOVE_TYPE.INVENTORY_TO_STATION
+      || task.move_type === MOVE_TYPE.STATION_TO_STATION
+      || task.move_type === MOVE_TYPE.WAITING_LOCATION_TO_STATION
+    ){
+      const gtpLocations = await this.gtpLocationRepository.find({where: {station_id: task.end_location.location_id}});
+      if (gtpLocations.length === 0){
+        throw new NotFoundException(`No GTP Locations found for station ID ${task.end_location.location_id}`);
+      }
+      const orderItems = await this.orderItemRepository.find({
+        where: {
+          source_location_id: task.origin_location,
+          status: OrderItemStatus.IN_PROGRESS,
+          destination_pallet_slot_id: In(gtpLocations.map(loc => loc.gtp_location_id)),
+        },
+      });
+      if (orderItems.length === 0){
+        throw new NotFoundException(`No Order Items found in PROGRESS for source location ID ${task.origin_location} at station ID ${task.end_location.location_id}`);
+      }
+    }
     return await this.orchestrationService.handleErroneousTask(taskId);
+  }
+
+  async retryOrderItem(orderItemId: number) {
+    const orderItem = await this.orderItemRepository.findOne({ where: { order_item_id: orderItemId } });
+    if (!orderItem){
+      throw new NotFoundException(`Order item with ID ${orderItemId} not found`);
+    }
+    const cancelledOrderItems = await this.orderItemRepository.find({
+      where: { merged_order_item_id: orderItem.merged_order_item_id || orderItemId , status: OrderItemStatus.CANCELLED }
+    });
+
+    for (const item of cancelledOrderItems){
+      item.status = OrderItemStatus.IN_PROGRESS;
+      item.retry_reassign_attempts += 1;
+      await this.orderItemRepository.save(item);
+      await this.loggingService.log(`Order Item ID ${item.order_item_id} set to IN_PROGRESS for retry`,
+        TaskType.GOODS_TO_PERSON, null, item.order_batch_id || '');
+    }
+
+    orderItem.status = OrderItemStatus.IN_PROGRESS;
+    orderItem.retry_reassign_attempts += 1;
+    await this.orderItemRepository.save(orderItem);
+    await this.loggingService.log(`Order Item ID ${orderItem.order_item_id} set to IN_PROGRESS for retry`,
+      TaskType.GOODS_TO_PERSON, null, orderItem.order_batch_id || '');
+
+    const task = await this.taskRepository.findOne({
+      where: {
+        origin_location: orderItem.source_location_id,
+        status: TaskStatus.CANCELLED,
+      },
+      order: { created_at: 'DESC' }
+    });
+
+    if (task){
+      await this.orchestrationService.handleErroneousTask(task.task_id);
+    }
+    else{
+      const gtpLocations = await this.gtpLocationRepository.find({where: { gtp_location_id: orderItem.destination_pallet_slot_id }});
+      if (gtpLocations.length === 0){
+        throw new NotFoundException(`No GTP Locations found for location ID ${orderItem.destination_pallet_slot_id}`);
+      }
+      await this.productRequirementRepository.save({
+        source_location_id: orderItem.source_location_id,
+        station_id: gtpLocations[0].station_id,
+      });
+    }
   }
 }
